services/vision/pyproject.toml
[tool.poetry]
name = "vision"
version = "0.1.0"
package-mode = false

[tool.poetry.dependencies]
python = "3.11.9"
onnxruntime = "1.18.1"
onnx = "^1.16.0"
numpy = "^2.1.0"
opencv-python-headless = "^4.10.0"
pillow = "^10.4.0"
pyscreenshot = "^3.1"
grpcio = "^1.65.0"
protobuf = "^5.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
ruff = "^0.6.0"
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
